The snapshot function has been upgraded to use `nbs_count` instead of `bp.count`.
The `set_exposure` and `number_exposure` code was removed, as `nbs_count` already handles this.  In addition to this primary update, the message printed in the boxed text was simplified to only display the number of counts, seconds, and detector name, leaving out the sample name and energy.  The variables used to generate plural units also were removed.  The ability to change the sample name temporarily also was removed because the old code only made the update in `RE.md`.  While it is possible to propagate changes to the `nbs-bl` sample manager, this feature seems unnecessary.  Instead, a new sample can be created staightforwardly, and then `snapshot` can be run.